requirements.txt
alembic==1.5.3
aniso8601==8.1.1
attrs==20.3.0
bcrypt==3.2.0
black==23.11.0
boto3==1.29.2
botocore==1.32.2
cachetools==5.2.0
certifi==2022.6.15
cffi==1.14.4
charset-normalizer==2.1.0
click==8.1.7
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.6.0
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-Testing==0.8.1
google-api-core==2.8.2
google-api-python-client==2.53.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.3
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==1.0.1
jsonschema==3.2.0
Mako==1.1.4
MarkupSafe==1.1.1
mypy-extensions==1.0.0
oauth2client==4.1.3
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
protobuf==4.21.2
psycopg2-binary==2.9.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.1
pyparsing==3.0.9
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==1.0.0
python-editor==1.0.4
pytz==2020.5
requests==2.28.1
rsa==4.8
s3transfer==0.7.0
six==1.15.0
SQLAlchemy==1.3.22
tomli==2.0.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==1.26.10
Werkzeug==0.16.1
